move week and remove dateservice
