+++ backend/Api/Organisation/OrganisationHolidayExtensions.cs
@@ -1,5 +1,4 @@
 using Core.DomainModels;
-using Core.Services;
 using PublicHoliday;
 
 namespace Api.Organisation;
@@ -8,7 +7,7 @@ public static class OrganisationHolidayExtensions
 {
     public static int GetTotalHolidaysOfWeek(this Organization organization, Week week)
     {
-        var datesOfThisWeek = DateService.GetDatesInWorkWeek(week.Year, week.WeekNumber);
+        var datesOfThisWeek = week.GetDatesInWorkWeek();
         return datesOfThisWeek.Count(organization.IsHoliday);
     }
