+++ backend/Api/Staffing/ReadModelFactory.cs
@@ -1,7 +1,6 @@
 using Api.Common;
 using Api.Organisation;
 using Core.DomainModels;
-using Core.Services;
 
 namespace Api.Staffing;
 
@@ -16,8 +15,8 @@ public class ReadModelFactory
 
     public List<ConsultantReadModel> GetConsultantReadModelsForWeeks(string orgUrlKey, List<Week> weeks)
     {
-        var firstDayInScope = DateService.FirstDayOfWorkWeek(weeks.First());
-        var firstWorkDayOutOfScope = DateService.LastWorkDayOfWeek(weeks.Last()).AddDays(1);
+        var firstDayInScope = weeks.First().FirstDayOfWorkWeek();
+        var firstWorkDayOutOfScope = weeks.Last().LastWorkDayOfWeek().AddDays(1);
 
         return _storageService.LoadConsultants(orgUrlKey)
             .Where(c => c.EndDate == null || c.EndDate > firstDayInScope)
@@ -107,14 +106,14 @@ public class ReadModelFactory
         var detailedBookings = billableBookings.Concat(offeredBookings).Concat(plannedAbsencesPrWeek);
 
         var vacationsInSet =
-            consultant.Vacations.Where(v => weekSet.Any(week => DateService.DateIsInWeek(v.Date, week)))
+            consultant.Vacations.Where(v => weekSet.Any(week => week.ContainsDate(v.Date)))
                 .ToList();
 
         if (vacationsInSet.Count > 0)
         {
             var vacationsPrWeek = weekSet.Select(week => new WeeklyHours(
                 week.ToSortableInt(),
-                vacationsInSet.Count(vacation => DateService.DateIsInWeek(vacation.Date, week)) *
+                vacationsInSet.Count(vacation => week.ContainsDate(vacation.Date)) *
                 consultant.Department.Organization.HoursPerWorkday
             )).ToList();
             detailedBookings = detailedBookings.Append(new DetailedBooking(
@@ -174,8 +173,8 @@ public class ReadModelFactory
 
     private static string GetDatesForWeek(Week week)
     {
-        return DateService.GetDatesInWorkWeek(week.Year, week.WeekNumber)[0].ToString("dd.MM") +
-               " - " + DateService
-                   .GetDatesInWorkWeek(week.Year, week.WeekNumber)[^1].ToString("dd.MM");
+        return week.GetDatesInWorkWeek()[0].ToString("dd.MM") +
+               " - " + week
+                   .GetDatesInWorkWeek()[^1].ToString("dd.MM");
     }
 }
